imrunicorn/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.1
Django==3.0.7
django-crispy-forms==1.8.1
django-fontawesome==1.0
django-model-utils==4.0.0
djangorestframework==3.11.0
django-environ==0.4.5
environ==1.0
gunicorn==20.0.4
idna==2.8
pytz==2019.3
python-dateutil
python-decouple
PyYAML==5.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
# testing an upgrade of mysqlclient
mysqlclient==2.0.3
# mysqlclient==1.4.6 # doesn't work on windows but isnt needed
PyMySQL==0.9.3 # doesn't work on windows but isnt needed
django-ipware==2.1.0  # used to get remote ip
django-extensions
Pillow
django-phone-field==1.8.1
django-iprestrict==1.7.0
django-ip-restriction==1.1.1
django-createsuperuserwithpassword==2.0.0
glom
